refactoring: Pavage et scratchblock sortent de 2d.js fix: fix:
