+++ src/js/exercices/3e/3G12.js
@@ -2,8 +2,9 @@ import Exercice from '../Exercice.js'
 import { mathalea2d, colorToLatexOrHTML } from '../../modules/2dGeneralites.js'
 import { context } from '../../modules/context.js'
 import { listeQuestionsToContenu, egal, randint, choice, shuffle, nombreAvecEspace, arcenciel, texcolors } from '../../modules/outils.js'
-import { pavage, tracePoint, labelPoint, segment, rotation, codageAngle, texteParPosition } from '../../modules/2d.js'
+import { tracePoint, labelPoint, segment, rotation, codageAngle, texteParPosition } from '../../modules/2d.js'
 import { rotationAnimee } from '../../modules/2dAnimation.js'
+import { pavage } from '../../modules/Pavage'
 export const titre = 'Trouver l\'image d\'une figure par une rotation dans un pavage'
 
 /**
