+++ src/js/exercices/3e/3I10-1.js
@@ -1,11 +1,13 @@
 import Exercice from '../Exercice.js'
 import { choice, combinaisonListes, contraindreValeur, deuxColonnes, lampeMessage, lettreMinusculeDepuisChiffre, listeQuestionsToContenu, randint, texteGras } from '../../modules/outils.js'
-import { point, scratchblock } from '../../modules/2d.js'
+import { point } from '../../modules/2d.js'
 import { noteLaCouleur, plateau2dNLC } from '../../modules/noteLaCouleur.js'
 import { colorToLatexOrHTML, fixeBordures, mathalea2d } from '../../modules/2dGeneralites.js'
 import { ajouterAx, ajouterAy, allerA, angleScratchTo2d, attendre, baisseCrayon, creerLutin, leveCrayon, orienter } from '../../modules/2dLutin.js'
 import { context } from '../../modules/context.js'
 import { propositionsQcm } from '../../modules/interactif/questionQcm.js'
+import { scratchblock } from '../../modules/scratchblock'
+
 export const titre = 'Analyser des scripts Scratch'
 export const interactifReady = true
 export const interactifType = 'qcm'
