+++ src/js/exercices/4e/4G11.js
@@ -2,8 +2,9 @@ import Exercice from '../Exercice.js'
 import { mathalea2d, colorToLatexOrHTML } from '../../modules/2dGeneralites.js'
 import { context } from '../../modules/context.js'
 import { listeQuestionsToContenu, egal, randint, shuffle, nombreAvecEspace, texcolors } from '../../modules/outils.js'
-import { tracePoint, vecteur, segment, translation, codageSegment, texteParPosition, pavage } from '../../modules/2d.js'
+import { tracePoint, vecteur, segment, translation, codageSegment, texteParPosition } from '../../modules/2d.js'
 import { translationAnimee } from '../../modules/2dAnimation.js'
+import { pavage } from '../../modules/Pavage'
 export const titre = 'Trouver l\'image d\'une figure par une translation dans un pavage'
 
 /**
