+++ src/js/exercices/4e/4I1-2.js
@@ -4,8 +4,10 @@ import { mathalea2d, colorToLatexOrHTML } from '../../modules/2dGeneralites.js'
 import { context } from '../../modules/context.js'
 import { listeQuestionsToContenuSansNumero, randint, choice, calcul, shuffle, arrondi } from '../../modules/outils.js'
 // Ici ce sont les fonctions de la librairie maison 2d.js qui gèrent tout ce qui est graphique (SVG/tikz) et en particulier ce qui est lié à l'objet lutin
-import { scratchblock, tracePoint, point, segment, texteParPoint } from '../../modules/2d.js'
+import { tracePoint, point, segment, texteParPoint } from '../../modules/2d.js'
 import { allerA, angleScratchTo2d, avance, baisseCrayon, creerLutin, leveCrayon, orienter, tournerD, tournerG } from '../../modules/2dLutin.js'
+import { scratchblock } from '../../modules/scratchblock.js'
+
 export const interactifReady = true
 export const interactifType = 'cliqueFigure'
 export const amcReady = true
