+++ src/js/exercices/6e/6I11.js
@@ -2,10 +2,12 @@
 import Exercice from '../Exercice.js'
 import { mathalea2d, colorToLatexOrHTML } from '../../modules/2dGeneralites.js'
 import { context } from '../../modules/context.js'
-import { scratchblock, point, texteParPositionEchelle } from '../../modules/2d.js'
+import { point, texteParPositionEchelle } from '../../modules/2d.js'
 import { allerA, angleScratchTo2d, attendre, baisseCrayon, clone, creerLutin, orienter } from '../../modules/2dLutin.js'
 import { noteLaCouleur, plateau2dNLC } from '../../modules/noteLaCouleur.js'
 import { choice, combinaisonListes, contraindreValeur, listeQuestionsToContenu, modalPdf, modalUrl, randint, stringNombre, texteGras } from '../../modules/outils.js'
+import { scratchblock } from '../../modules/scratchblock.js'
+
 export const titre = 'Note la couleur (scratch)'
 
 /**
