+++ src/js/exercices/c3/c3I11.js
@@ -3,7 +3,8 @@ import Exercice from '../Exercice.js'
 import { mathalea2d, colorToLatexOrHTML } from '../../modules/2dGeneralites.js'
 import { context } from '../../modules/context.js'
 import { randint, choice, texteGras, modalUrl, modalPdf, contraindreValeur, listeQuestionsToContenu, stringNombre } from '../../modules/outils.js'
-import { scratchblock, point, texteParPositionEchelle, tracePoint } from '../../modules/2d.js'
+import { point, texteParPositionEchelle, tracePoint } from '../../modules/2d.js'
+import { scratchblock } from '../../modules/scratchblock.js'
 import { noteLaCouleur, plateau2dNLC } from '../../modules/noteLaCouleur.js'
 import { allerA, angleScratchTo2d, attendre, baisseCrayon, clone, creerLutin, orienter } from '../../modules/2dLutin.js'
 export const titre = 'Note la couleur (scratch)'
