+++ src/js/exercices/can/6e/can6I01.js
@@ -3,10 +3,11 @@ import Exercice from '../../Exercice.js'
 import { mathalea2d, colorToLatexOrHTML, fixeBordures } from '../../../modules/2dGeneralites.js'
 import { context } from '../../../modules/context.js'
 import { randint, choice, texteGras, modalUrl, modalPdf, contraindreValeur, listeQuestionsToContenu, stringNombre } from '../../../modules/outils.js'
-import { scratchblock, point, texteParPositionEchelle, tracePoint, roseDesVents } from '../../../modules/2d.js'
+import { point, texteParPositionEchelle, tracePoint, roseDesVents } from '../../../modules/2d.js'
 import { noteLaCouleur, plateau2dNLC } from '../../../modules/noteLaCouleur.js'
 import { allerA, angleScratchTo2d, attendre, baisseCrayon, clone, creerLutin, orienter } from '../../../modules/2dLutin.js'
 import { ajouteChampTexte, setReponse } from '../../../modules/gestionInteractif.js'
+import { scratchblock } from '../../../modules/scratchblock.js'
 export const titre = 'Noter la couleur (scratch)'
 export const interactifType = 'mathLive'
 export const interactifReady = true
