+++ src/js/modules/2d.js
@@ -9948,383 +9948,6 @@ export function angleradian (A, O, B, precision = 2) {
 }
 
 /**
- * Traducteur scratch3 (Latex) -> scratchblocks
- * On lui passe une chaine de caractères contenant une série de commande Latex du package Latex Scratch3
- * Elle retourne une chaine de caractères contenant l'équivalent en langage scratchblocks si le contexte est isHtml !
- * Si le contexte est !isHtml alors elle retourne la chaine passée en argument.
- * http://mirrors.ctan.org/macros/latex/contrib/scratch3/scratch3-fr.pdf
- * https://scratchblocks.github.io
- * @author Jean-Claude Lhote.
- */
-
-export function scratchblock (stringLatex) {
-  const regex1 = /[\\{}]/
-  const regex3 = /[[]<>]/
-  const regex4 = /[{ ]/
-  const litcommande = function (souschaine) {
-    let extrait
-    if (souschaine[0] === '}') {
-      return '}'
-    } else {
-      extrait = souschaine.split(regex4)[0]
-      return extrait
-    }
-  }
-
-  /*****************************************************/
-  /** ********* La fonction d'analyse récursive *********/
-  /*****************************************************/
-  const translatex = function (chaine, index, compteAccolades) {
-    let resultat = []; let texte = []; let texte2 = []; let texte3 = []; let taille; let string; let fleche
-    let compteur, debut // pour les boucles et les if
-    const souschaine = chaine.substring(index)
-    const commande = litcommande(souschaine)
-    switch (commande.substring(0, 5)) {
-      case '\\bloc':
-        string = commande.split('{')[0]
-        taille = string.length
-        string = string.substring(6)
-        compteAccolades++
-        switch (string) {
-          case 'stop':
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            texte2 = translatex(chaine, texte[1], texte[2])
-            resultat = [`${texte[0]} ${texte2[0]} `, texte2[1], texte2[2]]
-            break
-          case 'move':
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            resultat = [texte[0], texte[1], texte[2]]
-            break
-          case 'variable':
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            resultat = [texte[0], texte[1], texte[2]]
-            break
-          case 'control':
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            resultat = [texte[0], texte[1], texte[2]]
-            break
-          case 'pen':
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            resultat = [texte[0] + ' :: pen', texte[1], texte[2]]
-            break
-          case 'list':
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            resultat = [texte[0] + ' :: list', texte[1], texte[2]]
-            break
-          case 'init':
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            resultat = [texte[0], texte[1], texte[2]]
-            break
-          case 'space\n':
-            compteAccolades--
-            resultat = ['\n', 11 + index, compteAccolades]
-            break
-          case 'if':
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            texte2 = translatex(chaine, texte[1], texte[2])
-            texte3 = translatex(chaine, texte2[1], texte2[2])
-            resultat = [`${texte[0]} ${texte2[0]} ${texte3[0]}`, texte3[1] + 1, texte3[2] - 1]
-            compteAccolades = resultat[2]
-            compteur = compteAccolades + 1
-            debut = chaine.substring(resultat[1]).indexOf('{') + resultat[1]
-            resultat[1] = debut + 1
-            resultat[0] += '\n'
-            while (compteur > compteAccolades) {
-              texte = translatex(chaine, resultat[1], compteur)
-              resultat[0] += ' ' + texte[0]
-              resultat[1] = texte[1]
-              compteur = texte[2]
-            }
-            resultat[0] += ' fin'
-            break
-
-          case 'ifelse':
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            texte2 = translatex(chaine, texte[1], texte[2])
-            texte3 = translatex(chaine, texte2[1], texte2[2])
-            resultat = [`${texte[0]} ${texte2[0]} ${texte3[0]}`, texte3[1] + 1, texte3[2] - 1]
-            compteAccolades = resultat[2]
-            compteur = compteAccolades + 1
-            debut = chaine.substring(resultat[1]).indexOf('{') + resultat[1]
-            resultat[1] = debut + 1
-            resultat[0] += '\n'
-            while (compteur > compteAccolades) {
-              texte = translatex(chaine, resultat[1], compteur)
-              resultat[0] += ' ' + texte[0]
-              resultat[1] = texte[1]
-              compteur = texte[2]
-            }
-            resultat[0] += ' sinon'
-            compteur = compteAccolades + 1
-            debut = chaine.substring(resultat[1]).indexOf('{') + resultat[1]
-            resultat[1] = debut + 1
-            resultat[0] += '\n'
-            while (compteur > compteAccolades) {
-              texte = translatex(chaine, resultat[1], compteur)
-              resultat[0] += ' ' + texte[0]
-              resultat[1] = texte[1]
-              compteur = texte[2]
-            }
-            resultat[0] += ' fin'
-            break
-          case 'repeat':
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            if (texte[0].split(' ')[1] !== 'indéfiniment') {
-              if (texte[0].split(' ')[1] !== "jusqu'à") {
-                texte2 = translatex(chaine, texte[1], texte[2])
-                texte3 = translatex(chaine, texte2[1], texte2[2])
-                resultat = [`${texte[0]} ${texte2[0]} ${texte3[0]}`, texte3[1] + 1, texte3[2] - 1]
-                compteAccolades = resultat[2]
-              } else {
-                texte2 = translatex(chaine, texte[1], texte[2])
-                resultat = [`${texte[0]} ${texte2[0]} `, texte2[1] + 1, texte2[2] - 1]
-                compteAccolades = resultat[2]
-              }
-            } else {
-              resultat = [`${texte[0]} `, texte[1] + 1, texte[2] - 1]
-              compteAccolades = resultat[2]
-            }
-            compteur = compteAccolades + 1
-            debut = chaine.substring(resultat[1]).indexOf('{') + resultat[1]
-            resultat[1] = debut + 1
-            resultat[0] += '\n'
-            while (compteur > compteAccolades) {
-              texte = translatex(chaine, resultat[1], compteur)
-              resultat[0] += ' ' + texte[0]
-              resultat[1] = texte[1]
-              compteur = texte[2]
-            }
-            resultat[0] += ' fin'
-            break
-          default:
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            resultat = [texte[0], texte[1], texte[2]]
-            break
-        }
-
-        break
-      case '\\oval':
-        string = commande.split('{')[0]
-        taille = string.length
-        string = string.substring(5)
-        compteAccolades++
-        if (string.charAt(string.length - 1) === '*') {
-          fleche = true
-          string = string.substring(0, string.length - 1)
-        } else fleche = false
-        switch (string) {
-          case 'num':
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            if (isNaN(texte[0]) && texte[0].indexOf(regex3)) {
-              resultat = [`[${texte[0]}]`, texte[1] + 1, texte[2] - 1]
-            } else {
-              resultat = [`(${texte[0]})`, texte[1] + 1, texte[2] - 1]
-            }
-            break
-          case 'moreblocks':
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            if (fleche) {
-              resultat = [`(${texte[0]} v)`, texte[1], texte[2]]
-            } else {
-              resultat = [`(${texte[0]})`, texte[1], texte[2]]
-            }
-            break
-          case 'variable':
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            if (fleche) {
-              resultat = [`(${texte[0]} v)`, texte[1] + 1, texte[2] - 1]
-            } else {
-              resultat = [`(${texte[0]})`, texte[1] + 1, texte[2] - 1]
-            }
-            break
-          case 'sound':
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            if (fleche) {
-              resultat = [`(${texte[0]} v :: sound)`, texte[1] + 1, texte[2] - 1]
-            } else {
-              resultat = [`(${texte[0]} :: sound)`, texte[1] + 1, texte[2] - 1]
-            }
-            break
-          case 'sensing':
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            if (fleche) {
-              resultat = [`(${texte[0]} v :: sensing)`, texte[1] + 1, texte[2] - 1]
-            } else {
-              resultat = [`(${texte[0]} :: sensing)`, texte[1] + 1, texte[2] - 1]
-            }
-            break
-          case 'operator':
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            texte2 = translatex(chaine, texte[1], texte[2])
-            resultat = [`(${texte[0]} ${texte2[0]}`, texte2[1], texte2[2]]
-            while (chaine.charAt(texte2[1]) !== '}') {
-              texte2 = translatex(chaine, texte2[1], texte2[2])
-              resultat[0] += ' ' + texte2[0]
-            }
-            resultat[0] += ')'
-            resultat[1] = texte2[1] + 1
-            resultat[2] = texte2[2] - 1
-            break
-
-          default:
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            if (fleche) {
-              resultat = [`(${texte[0]} v)`, texte[1] + 1, texte[2] - 1]
-            } else {
-              resultat = [`(${texte[0]})`, texte[1] + 1, texte[2] - 1]
-            }
-            break
-        }
-
-        break
-      case '\\bool':
-        string = commande.split(/\{ /)[0]
-        taille = string.length
-        string = string.substring(5, 9)
-        switch (string) {
-          case 'oper':
-            compteAccolades++
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            texte2 = translatex(chaine, texte[1], texte[2])
-            resultat = [`<${texte[0]} ${texte2[0]}`, texte2[1], texte2[2]]
-            while (chaine.charAt(texte2[1]) !== '}') {
-              texte2 = translatex(chaine, texte2[1], texte2[2])
-              resultat[0] += ' ' + texte2[0]
-            }
-            resultat[0] += ' :: operators boolean>'
-            resultat[1] = texte2[1] + 1
-            resultat[2] = texte2[2] - 1
-            break
-          case 'empt':
-            resultat = ['< vide :: operators boolean>', index + taille + 1, compteAccolades]
-            break
-          case 'sens':
-            compteAccolades++
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            texte2 = translatex(chaine, texte[1], texte[2])
-            resultat = [`<${texte[0]} ${texte2[0]}`, texte2[1], texte2[2]]
-            while (chaine.charAt(texte2[1]) !== '}') {
-              texte2 = translatex(chaine, texte2[1], texte2[2])
-              resultat[0] += ' ' + texte2[0]
-            }
-            resultat[0] += ' :: sensing>'
-            resultat[1] = texte2[1] + 1
-            resultat[2] = texte2[2] - 1
-            break
-          case 'list':
-            compteAccolades++
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            texte2 = translatex(chaine, texte[1], texte[2])
-            resultat = [`<${texte[0]} ${texte2[0]}`, texte2[1], texte2[2]]
-            while (chaine.charAt(texte2[1]) !== '}') {
-              texte2 = translatex(chaine, texte2[1], texte2[2])
-              resultat[0] += ' ' + texte2[0]
-            }
-            resultat[0] += ' :: list>'
-            resultat[1] = texte2[1] + 1
-            resultat[2] = texte2[2] - 1
-            break
-          default:
-            texte = translatex(chaine, index + taille + 1, compteAccolades)
-            resultat = [`<${texte[0]}>`, texte[1], texte[2]]
-            break
-        }
-        break
-      case '\\init':
-        string = commande.split('{')[0]
-        taille = string.length
-        compteAccolades++
-        texte = translatex(chaine, index + taille + 1, compteAccolades)
-        texte2 = translatex(chaine, texte[1], texte[2])
-        resultat = [`${texte[0]} ${texte2[0]} `, texte2[1], texte2[2]]
-        break
-      case '\\name':
-        string = commande.split('{')[0]
-        taille = string.length
-        compteAccolades++
-        texte = translatex(chaine, index + taille + 1, compteAccolades)
-        texte2 = translatex(chaine, texte[1], texte[2])
-        resultat = [`${texte[0]} ${texte2[0]}`, texte2[1], texte2[2]]
-        while (chaine.charAt(texte2[1]) !== '}') {
-          texte2 = translatex(chaine, texte2[1], texte2[2])
-          resultat[0] += ' ' + texte2[0]
-        }
-        resultat[1] = texte2[1] + 1
-        resultat[2] = texte2[2] - 1
-        break
-      default:
-        switch (commande) {
-          case '}':
-            compteAccolades--
-            resultat = [' ', 1 + index, compteAccolades]
-            break
-          case '\\begin':
-            compteAccolades++
-            if (chaine.substring(15 + index)[0] === '[') {
-              index = chaine.substring(15 + index).indexOf(']') + 16 + index
-            } else {
-              index += 15
-            }
-            resultat = [' <!-- Code Scratch  -->', index, compteAccolades]
-            break
-          case '\\end':
-            compteAccolades--
-            resultat = [' <!-- Fin du Code Scratch  -->\n', 13 + index, compteAccolades]
-            break
-          case '\\turnleft':
-            resultat = ['gauche ', 11 + index, compteAccolades]
-            break
-          case '\\turnright':
-            resultat = ['droite ', 12 + index, compteAccolades]
-            break
-          case '\\greenflag':
-            resultat = [' @greenFlag ', 10 + index, compteAccolades]
-            break
-          case '\\selectmenu':
-            compteAccolades++
-            texte = translatex(chaine, 12 + index, compteAccolades)
-            resultat = [`[${texte[0]} v]`, texte[1] + 1, texte[2] - 1]
-            break
-          default:
-            string = chaine.substring(index).split(regex1)[0]
-            resultat = [string, string.length + index, compteAccolades]
-            break
-        }
-        break
-    }
-    return resultat
-  }
-  /*********************************************/
-  /** *********** Fin de translatex *************/
-  /*********************************************/
-
-  // boucle pricipale de scratchblock2
-  let codeScratch = ''
-  let fin; let result = []; let index
-  let compteur = 0
-  if (!((stringLatex.match(/\{/g) || []).length === (stringLatex.match(/\}/g) || []).length)) {
-    // console.log("Il n'y a pas le même nombre de { que de }. Je préfère m'arrêter.")
-    return false
-  }
-  if (!context.isHtml) {
-    codeScratch = stringLatex
-  } else {
-    codeScratch = '<pre class=\'blocks\'>'
-    index = 0
-    fin = false
-    while (!fin) {
-      result = translatex(stringLatex, index, compteur)
-      codeScratch += result[0]
-      index = result[1]
-      compteur = result[2]
-      if (compteur === 0) fin = true
-    }
-    codeScratch += '</pre>\n'
-  }
-  return codeScratch
-}
-
-/**
  * Représente une boussole pour les 4 orientations possible dans Scratch
  * Il n'y a pas d'arguments, c'est un objet statique qui a sa méthode svg() et sa méthode tikz()
  * Il est difficile de modifier la taille de l'objet à cause du texte des blocks
@@ -11016,519 +10639,6 @@ export function Labyrinthe ({ nbLignes = 3, nbColonnes = 6, scaleFigure = 1 } = 
 // JSDOC Validee par EE Septembre 2022
 export function labyrinthe ({ nbLignes = 3, nbColonnes = 6, scaleFigure = 1 } = {}) {
   return new Labyrinthe({ nbLignes, nbColonnes, scaleFigure })
-}
-
-/**
- * Classe Pavage : permet de créer des pavages de polygones en un tour de main et de manipuler les polygones qu'il contient
- * @author Jean-Claude Lhote
- * publié le 10/12/2020
- */
-export function Pavage () {
-  this.type = 1
-  this.polygones = []
-  this.barycentres = []
-  this.tracesCentres = []
-  this.numeros = []
-  this.coordonnees = []
-  this.Nx = 1
-  this.Ny = 1
-  this.echelle = 20
-  this.fenetre = {}
-
-  this.construit = function (type = 1, Nx = 1, Ny = 1, taille = 3) {
-    const nettoieObjets = function (objets) {
-      let barywhite, baryblack // c'est drôle non ?
-      for (let i = 0; i < objets.length; i++) {
-        barywhite = barycentre(objets[i])
-        for (let j = i + 1; j < objets.length;) {
-          baryblack = barycentre(objets[j])
-          if (egal(barywhite.x, baryblack.x, 0.1) && egal(barywhite.y, baryblack.y, 0.1)) {
-            objets.splice(j, 1)
-          } else j++
-        }
-      }
-    }
-    let A; let B; let v; let w; let C; let D; let XMIN = 0; let YMIN = 0; let XMAX = 0; let YMAX = 0; let P1; let P2; let P3; let P4; let P5; let P6; let P7; let P8; let P9; let P10; let P11; let P12
-    A = point(0, 0)
-    B = point(taille, 0)
-    switch (type) {
-      case 1: // triangles équilatéraux
-        v = vecteur(A, B)
-        w = rotation(v, A, -90)
-        w = homothetie(w, A, 1.73205)
-        for (let k = 0; k < Ny; k++) {
-          for (let j = 0; j < Nx; j++) {
-            P1 = polygoneRegulier(A, B, 3)
-            P2 = rotation(P1, A, 60)
-            P3 = rotation(P1, A, -60)
-            P4 = rotation(P1, A, -120)
-            this.polygones.push(P1, P2, P3, P4)
-            for (const p of P1.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P2.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P3.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P4.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            A = translation(A, v)
-            B = translation(B, v)
-          }
-          A = translation(A, vecteur(-Nx * v.x, -2 * v.y))
-          B = translation(B, vecteur(-Nx * v.x, -2 * v.y))
-          A = translation(A, w)
-          B = translation(B, w)
-        }
-        break
-
-      case 2: // carrés
-        v = vecteur(A, B)
-        v = homothetie(v, A, 2)
-        w = rotation(v, A, -90)
-        for (let k = 0; k < Ny; k++) {
-          for (let j = 0; j < Nx; j++) {
-            P1 = polygoneRegulier(A, B, 4)
-            P2 = rotation(P1, A, 90)
-            P3 = rotation(P1, A, -90)
-            P4 = rotation(P1, A, -180)
-            this.polygones.push(P1, P2, P3, P4)
-
-            for (const p of P1.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P2.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P3.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P4.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            A = translation(A, v)
-            B = translation(B, v)
-          }
-          A = translation(A, vecteur(-Nx * v.x, -2 * v.y))
-          B = translation(B, vecteur(-Nx * v.x, -2 * v.y))
-          A = translation(A, w)
-          B = translation(B, w)
-        }
-        break
-
-      case 3: // hexagones
-        B = homothetie(B, A, 0.8)
-        v = vecteur(A, B)
-        v = homothetie(v, A, 2)
-        w = rotation(v, A, -90)
-        w = homothetie(w, A, 1.73205)
-        for (let k = 0; k < Ny; k++) {
-          for (let j = 0; j < Nx; j++) {
-            C = similitude(B, A, 30, 1.1547)
-            P1 = polygoneRegulier(A, C, 6)
-            P2 = rotation(P1, A, -120)
-            P3 = translation(P1, v)
-            P4 = translation(P2, v)
-            this.polygones.push(P1, P2, P3, P4)
-
-            for (const p of P1.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P2.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P3.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P4.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            A = translation(A, vecteur(2 * v.x, 0))
-            B = translation(B, vecteur(2 * v.x, 0))
-          }
-          A = translation(A, vecteur(-Nx * 2 * v.x, w.y))
-          B = translation(B, vecteur(-Nx * 2 * v.x, w.y))
-        }
-        break
-
-      case 4: // Pavage 3².4.3.4
-        v = vecteur(A, B)
-        v = homothetie(v, A, 2.73205)
-        w = rotation(v, A, -90)
-        for (let k = 0; k < Ny; k++) {
-          for (let j = 0; j < Nx; j++) {
-            C = rotation(B, A, 60)
-            P1 = polygoneRegulier(A, B, 3)
-            P2 = rotation(P1, A, 150)
-            P6 = rotation(P1, B, -150)
-            P7 = rotation(P1, B, 60)
-            P9 = rotation(P2, C, 150)
-            P10 = rotation(P9, A, -60)
-            P11 = rotation(P2, B, 60)
-            P12 = rotation(P6, A, -60)
-            P3 = polygoneRegulier(A, C, 4)
-            P4 = polygoneRegulier(C, B, 4)
-            P5 = rotation(P4, B, -150)
-            P8 = rotation(P3, A, 150)
-
-            this.polygones.push(P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12)
-
-            for (const p of P1.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P2.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P11.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-
-            for (const p of P12.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P3.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P4.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P5.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P6.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P7.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P8.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P9.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P10.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            A = translation(A, vecteur(v.x, 0))
-            B = translation(B, vecteur(v.x, 0))
-          }
-          A = translation(A, vecteur(-Nx * v.x, w.y))
-          B = translation(B, vecteur(-Nx * v.x, w.y))
-        }
-        break
-      case 5: // 4.8²
-        v = vecteur(A, B)
-        v = homothetie(v, A, 2.4142)
-        w = rotation(v, A, -90)
-
-        for (let k = 0; k < Ny; k++) {
-          for (let j = 0; j < Nx; j++) {
-            C = rotation(A, B, -135)
-            P1 = polygoneRegulier(A, B, 8)
-            P2 = polygoneRegulier(B, A, 8)
-            P3 = translation(P1, v)
-            P4 = translation(P2, v)
-            P5 = polygoneRegulier(C, B, 4)
-            P6 = translation(P5, v)
-            P7 = translation(P5, w)
-            P8 = translation(P6, w)
-            this.polygones.push(P1, P2, P3, P4, P5, P6, P7, P8)
-
-            for (const p of P1.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P2.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P3.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P4.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P5.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P6.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P7.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P8.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-
-            A = translation(A, vecteur(2 * v.x, 0))
-            B = translation(B, vecteur(2 * v.x, 0))
-          }
-          A = translation(A, vecteur(-Nx * 2 * v.x, 2 * w.y))
-          B = translation(B, vecteur(-Nx * 2 * v.x, 2 * w.y))
-        }
-        break
-
-      case 6: // Pavage hexagonal d'écolier
-        v = vecteur(A, B)
-        w = rotation(v, A, 60)
-        v = vecteur(v.x + w.x, v.y + w.y) // v=AB+CB
-        w = rotation(v, A, -60)
-
-        for (let k = 0; k < Ny; k++) {
-          for (let j = 0; j < Nx; j++) {
-            C = rotation(A, B, 120)
-            D = rotation(B, C, 60)
-            P1 = polygone(A, B, C, D)
-            P2 = rotation(P1, C, -60)
-            P3 = rotation(P1, A, 60)
-            P4 = translation(P2, v)
-            P5 = translation(P1, v)
-            P6 = translation(P3, v)
-            P7 = translation(P1, w)
-            P8 = translation(P2, w)
-            P9 = translation(P3, w)
-            this.polygones.push(P1, P2, P3, P4, P5, P6, P7, P8, P9)
-
-            for (const p of P1.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P2.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P3.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P4.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P5.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P6.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P7.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P8.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P9.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            A = translation(A, vecteur(w.x + v.x, w.y + v.y))
-            B = translation(B, vecteur(w.x + v.x, w.y + v.y))
-          }
-          A = translation(A, vecteur(-Nx * (w.x + v.x) + 2 * w.x - v.x, 2 * w.y - v.y))
-          B = translation(B, vecteur(-Nx * (w.x + v.x) + 2 * w.x - v.x, 2 * w.y - v.y))
-        }
-        break
-      case 7:
-        v = vecteur(A, B)
-        v = homothetie(v, A, 2)
-        w = rotation(v, A, -60)
-
-        for (let k = 0; k < Ny; k++) {
-          for (let j = 0; j < Nx; j++) {
-            C = rotation(A, B, -120)
-            D = rotation(B, C, -120)
-            P1 = polygoneRegulier(A, B, 6)
-            P2 = polygoneRegulier(C, B, 3)
-            P3 = rotation(P2, C, 180)
-            P4 = translation(P3, w)
-            P5 = translation(P2, w)
-            P6 = rotation(P1, B, 180)
-            this.polygones.push(P1, P2, P3, P6, P5, P4)
-
-            for (const p of P1.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P2.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P3.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P4.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P5.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            for (const p of P6.listePoints) {
-              XMIN = Math.min(XMIN, p.x)
-              XMAX = Math.max(XMAX, p.x)
-              YMIN = Math.min(YMIN, p.y)
-              YMAX = Math.max(YMAX, p.y)
-            }
-            A = translation(A, v)
-            B = translation(B, v)
-          }
-          A = translation(A, vecteur(-Nx * v.x + 2 * w.x - v.x, 2 * w.y - v.y))
-          B = translation(B, vecteur(-Nx * v.x + 2 * w.x - v.x, 2 * w.y - v.y))
-        }
-        break
-    }
-    this.echelle = 80 / Math.sqrt(XMAX - XMIN)
-    this.fenetre = { xmin: XMIN - 0.5, ymin: YMIN - 0.5, xmax: XMAX + 0.5, ymax: YMAX + 0.5, pixelsParCm: this.echelle, scale: this.echelle / 30 }
-    nettoieObjets(this.polygones) // On supprime les doublons éventuels (grâce à leur barycentre)
-    // On ajoute les N°
-    this.nb_polygones = this.polygones.length // Le nombre de polygones du pavage qui sert dans les boucles
-
-    for (let i = 0; i < this.nb_polygones; i++) {
-      this.barycentres.push(barycentre(this.polygones[i]))
-      this.tracesCentres.push(tracePoint(this.barycentres[i], 'blue'))
-      this.tracesCentres[i].opacite = 0.5
-      this.tracesCentres[i].taille = 2
-      this.coordonnees.push([this.barycentres[i].x, this.barycentres[i].y])
-      this.numeros.push(texteParPosition(nombreAvecEspace(i + 1), this.barycentres[i].x + 0.5, this.barycentres[i].y, 'milieu', 'black', 50 / this.echelle, 0, true))
-    }
-  }
-}
-export function pavage () {
-  return new Pavage()
 }
 
 /**
